[Sonic Origins] Various Additions and Fixes (#107)
* [Sonic Origins] Fix Enable Lives Codes in Sonic 1

Due to an incorrect global variable offset for `game.playMode`, the Enable Lives Code behaved incorrectly in Sonic 1. This fixes the offset, as well as applying the Sonic 1/2 game check seen in the S1/S2 game option codes.

* [Sonic Origins] Add GetCurrentGame() to RSDK Library

* [Sonic Origins] Add GetEngineVersion() to RSDK Library

* [Sonic Origins] Set LibraryInitializer attribute for RSDK Init()

* [Sonic Origins] Rework GetCurrentGame() to check Data filepath

I had to replace the Helpers library for Origins with a copy of the one for Forces to get this to work. I'm not sure if this would cause any issues, but it works from what I tested. Maybe this library could be reworked into a Global one?

* [Sonic Origins] Include Helpers in RSDK

* [Globals] Added Helpers library

Revert "[Globals] Added Helpers library"

This reverts commit 821fc539aa24cd5c9c970141a1f96ab9f407f96c.

[Globals] Added Helpers and StringMapOperation libraries

* [Sonic Origins] Rework S1/S2 global variable codes

Instead of constantly setting the S1/S2 global variables through separate codes, a bridge library is now used to handle which variables are set in an unmanaged function called when the game loads. This fixes the initial stage load issue with the shield/item codes.

* [Sonic Origins] Set variable normally in Disable Super Music

* [Sonic Origins] Change how GetCurrentGame() works again

The function now checks the engine version like before, but still checks the data path for S1 vs S2.

* [Sonic Origins] Enable Blue Sphere Options in Title Screen

* [Sonic Origins] Fix GetRSDKGlobalsPtr(), remove GetRSDKGlobalsOffset() and GetRSDKGlobalsAbsPtr()

* [Sonic Origins] Add PlayMode, S3KMedalMods, and S3KSecrets enums

* [Sonic Origins] Enable Super Peel Out

* [Sonic Origins] More Medal Mod/Secret codes

* [Sonic Origins] Add GetPlayMode() to RSDK Library

* Remove extra period

* [Sonic Origins] Add legacy global variable structs to RSDK library

I didn't add a struct for S3&K since I don't believe there's a full, updated list of the globals anywhere.